bitset thingy

+++ template/short_things/bitset_operation.cpp
@@ -1,3 +1,5 @@
+// not working in my pc
+
 // https://gist.github.com/cgiosy/a441de545c9e96b1d7b02cc7a00561f9
 
 #define private public
@@ -7,7 +9,7 @@
 
 template<size_t _Nw> void _M_do_sub(_Base_bitset<_Nw>& A, const _Base_bitset<_Nw>& B) {
   for(int i = 0, c = 0;i < _Nw;i++)
-    c = _subborrow_u64(c, A._M_w[i], B._M_w[i], (unsigned long long*) & A._M_w[i]);
+    c = _subborrow_u64(c, A._M_w[i], B._M_w[i], (ull*) & A._M_w[i]);
 }
 template<> void _M_do_sub(_Base_bitset<1>& A, const _Base_bitset<1>& B) {
   A._M_w -= B._M_w;
@@ -24,7 +26,7 @@ template<size_t _Nb> inline bitset<_Nb> operator-(const bitset<_Nb>& A, const bi
 
 template<size_t _Nw> void _M_do_add(_Base_bitset<_Nw>& A, const _Base_bitset<_Nw>& B) {
   for(int i = 0, c = 0;i < _Nw;i++)
-    c = _addcarry_u64(c, A._M_w[i], B._M_w[i], (unsigned long long*) & A._M_w[i]);
+    c = _addcarry_u64(c, A._M_w[i], B._M_w[i], (ull*) & A._M_w[i]);
 }
 template<> void _M_do_add(_Base_bitset<1>& A, const _Base_bitset<1>& B) {
   A._M_w += B._M_w;
